requirements.txt
asgiref==3.7.2
bootstrap4==0.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
confusable-homoglyphs==3.2.0
crispy-bulma==0.10.0
defusedxml==0.7.1
dj-database-url==2.1.0
Django==3.2.9
django-allauth==0.41.0
django-appconf==1.0.5
django-autocomplete-light==3.9.7
django-bower==5.2.0
django-crispy-forms==2.0
django-on-heroku==1.1.2
django-registration==3.3
django-registration-redux==2.4
django-rest-auth==0.9.5
django-select2==6.1.0
django-tables2==2.4.1
django-tables2-column-shifter==2.0.3
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
idna==3.4
isodate==0.6.1
oauthlib==3.2.2
Pillow==10.0.1
psycopg2-binary==2.9.8
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.5
whitenoise==6.5.0
